Telephony: Fix compilation issues for conversion from long to int.
- Compilation issues fix.

+++ src/com/android/incallui/AnswerPresenter.java
@@ -49,7 +49,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
             processVideoUpgradeRequestCall(call);
         }
         for (int i = 0; i < CallList.PHONE_COUNT; i++) {
-            long[] subId = CallList.getInstance().getSubId(i);
+            int[] subId = CallList.getInstance().getSubId(i);
             call = calls.getCallWithState(Call.State.INCOMING, 0, subId[0]);
             if (call == null) {
                 call = calls.getCallWithState(Call.State.CALL_WAITING, 0, subId[0]);
@@ -93,7 +93,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
 
     @Override
     public void onIncomingCall(Call call) {
-        long subId = call.getSubId();
+        int subId = call.getSubId();
         int phoneId = CallList.getInstance().getPhoneId(subId);
         // TODO: Ui is being destroyed when the fragment detaches.  Need clean up step to stop
         // getting updates here.
@@ -143,7 +143,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     }
 
     private void processIncomingCall(Call call) {
-        long subId = call.getSubId();
+        int subId = call.getSubId();
         int phoneId = CallList.getInstance().getPhoneId(subId);
         mCallId[phoneId] = call.getId();
         mCall[phoneId] = call;
@@ -167,7 +167,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
             return;
         }
 
-        long subId = call.getSubId();
+        int subId = call.getSubId();
         int phoneId = CallList.getInstance().getPhoneId(subId);
         mCallId[phoneId] = call.getId();
         mCall[phoneId] = call;
@@ -211,7 +211,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     public void onCallChanged(Call call) {
         Log.d(this, "onCallStateChange() " + call + " " + this);
         if (call.getState() != Call.State.INCOMING) {
-            long subId = call.getSubId();
+            int subId = call.getSubId();
             int phoneId = CallList.getInstance().getPhoneId(subId);
 
             boolean isUpgradePending = isVideoUpgradePending(call);
@@ -243,7 +243,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     private int getActivePhoneId() {
         int phoneId = -1;
         if (CallList.getInstance().isDsdaEnabled()) {
-            long subId = CallList.getInstance().getActiveSubscription();
+            int subId = CallList.getInstance().getActiveSubscription();
             phoneId = CallList.getInstance().getPhoneId(subId);
         } else {
             for (int i = 0; i < mCall.length; i++) {
@@ -372,7 +372,7 @@ public class AnswerPresenter extends Presenter<AnswerPresenter.AnswerUi>
     }
 
     @Override
-    public void onActiveSubChanged(long subId) {
+    public void onActiveSubChanged(int subId) {
         final CallList calls = CallList.getInstance();
         final Call call = calls.getIncomingCall();
         int phoneId = CallList.getInstance().getPhoneId(subId);
